Moving neighboursearch tests into test_interaction.py
